+++ Pipfile
@@ -5,6 +5,7 @@ name = "pypi"
 
 [packages]
 numpy = "1.26.4"
+pytest = "*"
 
 [dev-packages]
 ipykernel = "*"
