+++ src/smart_building_rating_calculator/inputs.py
